+++ src/simulation/structs.jl
@@ -15,6 +15,7 @@ struct ResponseType
     id::String
     static_coefficient_functions::SVector{NUM_COEFFICIENTS,Function} # each takes imprinted, matured sequences and returns Float64 coefficient
     specific_coefficient_functions::SVector{NUM_COEFFICIENTS,Function} # each takes imprinted, matured, and infecting sequences and returns Float64 coefficient
+    immunityProbability::Function # each takes imprinted, matured, and infecting sequences and returns Float64 coefficient
 end
 
 struct ClassParameters